Sciunit process tree created separately
The Sciunit process tree is now not created as a child process of the signal handler. Rather, it is created as its own process group by using start_new_session=True which calles setsid internally. This allows the Sciunit process tree to continue wrapping its auditing even when the Jupyter server sends a kill signal to the signal handler.